fix: minor restructuring
