+++ src/services/ApiService.unit.handlers.ts
@@ -1,4 +1,3 @@
-import { buildStepObject, mockTransactionRequest } from './../../test/fixtures'
 import { ChainId, CoinKey, findDefaultToken } from '@lifi/types'
 import { rest } from 'msw'
 import ConfigService from './ConfigService'
@@ -49,14 +48,7 @@ export const handlers = [
   rest.post(
     `${config.apiUrl}/advanced/stepTransaction`,
     async (request, response, context) => {
-      return response(
-        context.json({
-          ...buildStepObject({
-            includingExecution: true,
-          }),
-          transactionRequest: mockTransactionRequest,
-        })
-      )
+      return response(context.json({}))
     }
   ),
   rest.get(
